Include directories in sync
Previously directories were only created as needed for files, and
removed directories would not actually be removed. This would be fine
for git as it only tracks files anyway, so directories without files
will disappear automatically, but it's not good for other systems (and
it is a bit of pain for the tests as well).