Handle non-consecutive log files
In my testing the log file always seemed to have a number 1 higher than
the manifest, so I built the version logic assuming this would always be
the case.

It appears this is not actually true though, so this eventually broke.
The logic has now been adjusted to look for a single file ending in
`.log` in the same directory. The number of this file is now also part
of the version string (in addition to its size, which was already in
these).